Refactored views into their own files
